web/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-chess
pydantic>=2.0
python-multipart
